feat(journey-maps): move leg ID selection logic into map-route-service.ts
Refs: ROKAS-1204
